refactor(core): extract createActivationKeyHandlers
Pull the 4 keyboard activation handlers (handleEnterKeyActivation,
handleOpenFileShortcut, handleContextMenuKey, handleActivationKeys)
out of init() into a focused controller. They are ~200 lines of
densely-coupled keyboard logic that coordinate the activation/hold
state machine, the prompt-mode entry path, the open-file shortcut,
the keyboard context-menu invocation, and the keydown-spam auto-
deactivation.

The new module is named activation-key-handlers.ts to avoid colliding
with the pre-existing keyboard-handlers.ts (setupKeyboardEventClaimer
/ getRequiredModifiers).

Removes 7 now-unused imports (hasTextSelectionInInput,
hasTextSelectionOnPage, KEYDOWN_SPAM_TIMEOUT_MS,
INPUT_FOCUS_ACTIVATION_DELAY_MS, INPUT_TEXT_SELECTION_ACTIVATION_DELAY_MS,
openFile, activateRenderer destructure).

index.tsx: 1726 -> 1546 (-180). 97 keyboard+activation+open-file specs
green.
